+++ src/main/java/com/wanted/preonboarding/ticket/domain/entity/PerformanceId.java
@@ -1,4 +1,4 @@
-package com.wanted.preonboarding.ticket.domain;
+package com.wanted.preonboarding.ticket.domain.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
