--- src/main/java/com/wanted/preonboarding/ticket/domain/infrastructure/PerformanceRepository.java
@@ -1,8 +1,0 @@
-package com.wanted.preonboarding.ticket.domain.infrastructure;
-
-import com.wanted.preonboarding.ticket.domain.Performance;
-import com.wanted.preonboarding.ticket.domain.PerformanceId;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PerformanceRepository extends JpaRepository<Performance, PerformanceId> {
-}
