+++ src/main/java/com/wanted/preonboarding/ticket/infrastructure/PerformanceRepository.java
@@ -1,0 +1,8 @@
+package com.wanted.preonboarding.ticket.infrastructure;
+
+import com.wanted.preonboarding.ticket.domain.entity.Performance;
+import com.wanted.preonboarding.ticket.domain.entity.PerformanceId;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PerformanceRepository extends JpaRepository<Performance, PerformanceId> {
+}
