+++ src/main/java/com/wanted/preonboarding/ticket/interfaces/controller/TicketController.java
@@ -1,8 +1,8 @@
 package com.wanted.preonboarding.ticket.interfaces.controller;
 
+import com.wanted.preonboarding.ticket.domain.service.ReservationService;
 import com.wanted.preonboarding.ticket.interfaces.dto.*;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -16,7 +16,6 @@ import java.util.List;
 public class TicketController {
 
     private final ReservationService reservationService;
-
 
     @PostMapping(path = "/reservation")
     public ResultResponse<ReservationResponseDto> reserve(ReservationRequestDto reservationRequestDto) {
